requirements.txt
openai>=1.12.0
openai-agents
requests>=2.31.0
python-dotenv>=1.0.0
replicate>=0.23.0
